requirements.txt
Django==4.1
PyMySQL==1.1.1
python-dotenv==1.0.1
pyTelegramBotAPI==4.22.1
requests==2.32.3
python-crontab==3.2.0
django-crontab==0.7.1
python-docx==1.1.2
docxtpl==0.19.1
openai==1.51.0
httpx==0.27.2
setuptools<81.0.0
reportlab==4.1.0
python-docx-replace==0.4.4
PyPDF2==3.0.1
striprtf==0.0.26